+++ kp_326/kp_tasks/farm_association/server/src/handling/FarmPlotAssignment.cpp
@@ -46,12 +46,12 @@ fasc::server::controllers::dto::FarmPlotAssignmentKeyDto keyFrom(
     const fasc::server::core::HttpRequest& request) {
   fasc::server::controllers::dto::FarmPlotAssignmentKeyDto key;
   if (const auto it = request.query_params.find("farm_id"); it != request.query_params.end()) {
-    key.farmId = std::stoi(it->second);
+    key.farmId = std::stoull(it->second);
   } else {
     throw std::invalid_argument{"Missing key field: farm_id"};
   }
   if (const auto it = request.query_params.find("farm_plot_id"); it != request.query_params.end()) {
-    key.farmPlotId = std::stoi(it->second);
+    key.farmPlotId = std::stoull(it->second);
   } else {
     throw std::invalid_argument{"Missing key field: farm_plot_id"};
   }
